+++ PAC_control_projects/main_control_prg/main_cycle.cpp
@@ -89,11 +89,11 @@ int main_cycle()
     timeInfo_ = get_time();
     static int print_cycle_last_h = timeInfo_.tm_hour;
 
-    static u_long max_iteration_cycle_time = 0;
+    static uint32_t max_iteration_cycle_time = 0;
     static u_int cycles_per_period = 0;
     cycles_per_period++;
 
-    static u_long cycle_time = 0;
+    static uint32_t cycle_time = 0;
     cycle_time = get_delta_millisec( st_time );
     G_PAC_INFO()->set_cycle_time( cycle_time );
 
@@ -105,7 +105,7 @@ int main_cycle()
     //Once per hour writing performance info.
     if ( print_cycle_last_h != timeInfo_.tm_hour )
         {
-        u_long avg_time = all_time / cycles_cnt;
+        uint32_t avg_time = all_time / cycles_cnt;
 
         if ( TRESH_AVG < avg_time )
             {
